+++ api/tests/tooling/domain-builder/factory/index.js
@@ -44,6 +44,7 @@ module.exports = {
   buildTube: require('./build-tube'),
   buildTutorial: require('./build-tutorial'),
   buildUser: require('./build-user'),
+  buildUserScorecard: require('./build-user-scorecard'),
   buildValidation: require('./build-validation'),
   buildValidator: require('./build-validator'),
 };