+++ mon-pix/app/models/user.js
@@ -16,6 +16,7 @@ export default Model.extend({
   organizations: hasMany('organization'),
   certifications: hasMany('certification'),
   campaignParticipations: hasMany('campaign-participation'),
+  scorecards: hasMany('scorecard'),
 
   competenceAreas: computed('competences', function() {
     return this.competences.then((competences) => {
@@ -36,5 +37,9 @@ export default Model.extend({
 
   fullName: computed('firstName', 'lastName', function() {
     return `${this.firstName} ${ this.lastName}`;
-  })
+  }),
+
+  areasCode: computed('scorecards.@each.area', function() {
+    return this.scorecards.mapBy('area.code').uniq();
+  }),
 });